Refactored payment page to be reusable

+++ src/app/components/login/login.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { AuthService } from 'src/app/services/auth/auth.service';
+import { SharedService } from 'src/app/services/shared/shared.service';
 
 @Component({
   selector: 'app-login',
@@ -14,7 +15,8 @@ export class LoginComponent implements OnInit {
   constructor(
     private authService: AuthService,
     private router: Router,
-  ) {}
+    private sharedService: SharedService,
+  ) { }
 
   ngOnInit() {
     this.logout();
@@ -24,28 +26,28 @@ export class LoginComponent implements OnInit {
     this.authService.authenticate(this.usernameOrEmail, this.password).subscribe(
       (response) => {
         this.quickCheck(response[0].id);
-        localStorage.setItem('username', response[0].username );
-        localStorage.setItem('user_id', response[0].id );
-        localStorage.setItem('is_authenticated', 'true' );
+        localStorage.setItem('username', response[0].username);
+        localStorage.setItem('user_id', response[0].id);
+        localStorage.setItem('is_authenticated', 'true');
         this.router.navigate(['/feeds']);
-      },
-      (error) => {
-        console.error('Authentication failed', error);
-        // Handle failed login
-      }
-    );
+        const currentUrl = this.router.url;
+        if(currentUrl == '/feeds') {
+          this.sharedService.refreshPage();
+        }
+      });
   }
 
   logout() {
-    // localStorage.removeItem('user_id');
-    // localStorage.removeItem('username');
-    // localStorage.removeItem('is_authenticated');
-    localStorage.clear();
-    this.router.navigate(['/login']);
+    localStorage.removeItem('user_id');
+    localStorage.removeItem('username');
+    localStorage.removeItem('is_authenticated');
+    // localStorage.removeItem('viewedPosts');
+    // localStorage.clear();
+    // this.router.navigate(['/login']);
   }
 
   quickCheck(user_id: string) {
-    if (localStorage.getItem('user_id') != user_id) {
+    if (localStorage.getItem('user_id') && localStorage.getItem('user_id') != user_id) {
       console.log('Clear all items');
       localStorage.clear();
     }
